[PSym] Add iterative symbolic search
Add back iterative symbolic search cleanly

Refactors code to push iterative search code to generic search scheduler

Keeping explicit/symbolic schedulers still separate

Updates logging based on search config
