+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/EntryPoint.java
@@ -24,8 +24,7 @@ import psym.valuesummary.solvers.SolverEngine;
 public class EntryPoint {
   private static ExecutorService executor;
   private static Future<Integer> future;
-  private static String status;
-  private static SearchScheduler scheduler;
+  private static SearchScheduler searchScheduler;
 
   private static void runWithTimeout(long timeLimit)
       throws TimeoutException,
@@ -60,16 +59,14 @@ public class EntryPoint {
 
   private static void print_stats() {
     double searchTime = TimeMonitor.getInstance().stopInterval();
-    TimeMonitor.getInstance().startInterval();
-    scheduler.print_search_stats();
-    if (scheduler.isFinalResult && scheduler.result.equals("correct for any depth")) {
-      status = "proved";
+    searchScheduler.print_search_stats();
+    if (searchScheduler.isFinalResult && PSymGlobal.getResult().equals("correct for any depth")) {
+      PSymGlobal.setStatus("proved");
     }
     StatWriter.log("time-search-seconds", String.format("%.1f", searchTime));
     if (PSymGlobal.getConfiguration().isIterative()) {
-      ((SearchScheduler) scheduler).reportEstimatedCoverage();
+      ((SearchScheduler) searchScheduler).reportEstimatedCoverage();
     }
-    StatWriter.log("status", String.format("%s", status));
   }
 
   private static void preprocess() {
@@ -81,9 +78,10 @@ public class EntryPoint {
     PSymLogger.info("--------------------");
 
     executor = Executors.newSingleThreadExecutor();
-    status = "error";
+
+    PSymGlobal.setResult("error");
     if (PSymGlobal.getConfiguration().getSymmetryMode() != SymmetryMode.None) {
-      SymmetryTracker.setScheduler(scheduler);
+      SymmetryTracker.setScheduler(searchScheduler);
     }
 
     double preSearchTime =
@@ -103,38 +101,41 @@ public class EntryPoint {
     if (printStats) {
       print_stats();
     }
-    if (scheduler instanceof ExplicitSearchScheduler) {
-      ExplicitSearchScheduler explicitSearchScheduler = (ExplicitSearchScheduler) scheduler;
+    if (searchScheduler instanceof ExplicitSearchScheduler) {
+      ExplicitSearchScheduler explicitSearchScheduler = (ExplicitSearchScheduler) searchScheduler;
       PSymLogger.finishedExplicit(
               explicitSearchScheduler.getIter(),
               explicitSearchScheduler.getIter() - explicitSearchScheduler.getStart_iter(),
               Duration.between(TimeMonitor.getInstance().getStart(), end).getSeconds(),
-              scheduler.result);
+              PSymGlobal.getResult());
     } else {
-      assert (scheduler instanceof SymbolicSearchScheduler);
+      assert (searchScheduler instanceof SymbolicSearchScheduler);
       PSymLogger.finishedSymbolic(
               Duration.between(TimeMonitor.getInstance().getStart(), end).getSeconds(),
-              scheduler.result);
+              PSymGlobal.getResult());
     }
   }
 
   private static void process(boolean resume) throws Exception {
     try {
-      TimedCall timedCall = new TimedCall(scheduler, resume);
+      TimedCall timedCall = new TimedCall(searchScheduler, resume);
       future = executor.submit(timedCall);
       TimeMonitor.getInstance().startInterval();
       runWithTimeout((long) PSymGlobal.getConfiguration().getTimeLimit());
-      status = "completed";
+      PSymGlobal.setStatus("completed");
     } catch (TimeoutException e) {
-      status = "timeout";
+      PSymGlobal.setStatus("timeout");
       throw new Exception("TIMEOUT", e);
     } catch (MemoutException | OutOfMemoryError e) {
-      status = "memout";
+      PSymGlobal.setStatus("memout");
       throw new Exception("MEMOUT", e);
     } catch (BugFoundException e) {
-      status = "cex";
-      scheduler.result = "found cex of length " + scheduler.getDepth();
-      scheduler.isFinalResult = true;
+      PSymGlobal.setStatus("cex");
+      PSymGlobal.setResult(String.format(
+              "found cex of %slength %d",
+              PSymGlobal.getConfiguration().isSymbolic() ? "symbolic " : "",
+              searchScheduler.getDepth()));
+      searchScheduler.isFinalResult = true;
       postprocess(true);
       PSymLogger.info(e.toString());
       if (PSymGlobal.getConfiguration().getVerbosity() > 0) {
@@ -149,33 +150,33 @@ public class EntryPoint {
 
       ReplayScheduler replayScheduler =
           new ReplayScheduler(
-              scheduler.getProgram(),
-              scheduler.getSchedule(),
+              searchScheduler.getProgram(),
+              searchScheduler.getSchedule(),
               pc);
       PSymGlobal.setScheduler(replayScheduler);
       replay(replayScheduler);
     } catch (InterruptedException e) {
-      status = "interrupted";
+      PSymGlobal.setStatus("interrupted");
       throw new Exception("INTERRUPTED", e);
     } catch (RuntimeException e) {
-      status = "error";
+      PSymGlobal.setStatus("error");
       throw new Exception("ERROR", e);
     } finally {
       //            GlobalData.getChoiceLearningStats().printQTable();
       future.cancel(true);
       executor.shutdownNow();
       TraceLogger.setVerbosity(0);
-      postprocess(!status.equals("cex"));
+      postprocess(!PSymGlobal.getStatus().equals("cex"));
     }
   }
 
   public static void run(Program p) throws Exception {
     if (PSymGlobal.getConfiguration().isSymbolic()) {
-      scheduler = new SymbolicSearchScheduler(p);
+      searchScheduler = new SymbolicSearchScheduler(p);
     } else {
-      scheduler = new ExplicitSearchScheduler(p);
+      searchScheduler = new ExplicitSearchScheduler(p);
     }
-    PSymGlobal.setScheduler(scheduler);
+    PSymGlobal.setScheduler(searchScheduler);
 
     preprocess();
     process(false);
@@ -183,8 +184,8 @@ public class EntryPoint {
 
   public static void resume() throws Exception {
     assert (PSymGlobal.getConfiguration().isExplicit());
-    scheduler = ExplicitSearchScheduler.readFromFile(PSymGlobal.getConfiguration().getReadFromFile());
-    PSymGlobal.setScheduler(scheduler);
+    searchScheduler = ExplicitSearchScheduler.readFromFile(PSymGlobal.getConfiguration().getReadFromFile());
+    PSymGlobal.setScheduler(searchScheduler);
     TraceLogger.setVerbosity(PSymGlobal.getConfiguration().getVerbosity());
     SolverEngine.resumeEngine();
 
@@ -198,13 +199,14 @@ public class EntryPoint {
       ScheduleWriter.Initialize();
       TextWriter.Initialize();
       replayScheduler.doSearch();
-      status = "error";
+      PSymGlobal.setStatus("error");
       throw new RuntimeException("ERROR: Failed to replay counterexample");
     } catch (BugFoundException e) {
-      if (replayScheduler.getCexLength() != replayScheduler.getChoiceDepth()) {
-        status = "error";
+      if (replayScheduler.getChoiceDepth() != replayScheduler.schedule.size()) {
+        PSymGlobal.setStatus("error");
         throw new RuntimeException("ERROR: Failed to replay counterexample");
       } else {
+        PSymGlobal.setResult(String.format("found cex of length %d", replayScheduler.getDepth()));
         PSymLogger.info(e.toString());
         PSymGlobal.printStackTrace(e, true);
         PSymLogger.info("Checker found a bug.");
@@ -237,6 +239,6 @@ public class EntryPoint {
 
   public static void writeToFile() throws Exception {
     assert (PSymGlobal.getConfiguration().isExplicit());
-    ((ExplicitSearchScheduler) scheduler).writeToFile();
+    ((ExplicitSearchScheduler) searchScheduler).writeToFile();
   }
 }
