+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/replay/ReplayScheduler.java
@@ -20,9 +20,6 @@ public class ReplayScheduler extends Scheduler {
   @Getter
   /** Path constraint */
   private final Guard pathConstraint;
-  @Getter
-  /** Counterexample length */
-  private final int cexLength;
 
   public ReplayScheduler(
       Program p,
@@ -35,7 +32,6 @@ public class ReplayScheduler extends Scheduler {
       machine.reset();
     }
     PSymGlobal.getConfiguration().setToReplay();
-    cexLength = this.schedule.size();
     pathConstraint = pc;
   }
 
@@ -61,7 +57,7 @@ public class ReplayScheduler extends Scheduler {
 
   @Override
   public void doSearch() throws TimeoutException {
-    TraceLogger.logStartReplayCex(cexLength);
+    TraceLogger.logStartReplayCex();
     ScheduleWriter.logHeader();
     TextWriter.logHeader();
     initializeSearch();
