+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/search/symbolic/SymbolicSearchScheduler.java
@@ -1,5 +1,6 @@
 package psym.runtime.scheduler.search.symbolic;
 
+import java.math.BigDecimal;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.*;
@@ -18,6 +19,7 @@ import psym.runtime.machine.events.Message;
 import psym.runtime.scheduler.search.SearchScheduler;
 import psym.runtime.scheduler.search.explicit.StateCachingMode;
 import psym.runtime.scheduler.search.symmetry.SymmetryMode;
+import psym.runtime.statistics.CoverageStats;
 import psym.runtime.statistics.SearchStats;
 import psym.runtime.statistics.SolverStats;
 import psym.utils.Assert;
@@ -30,66 +32,38 @@ import psym.valuesummary.ValueSummary;
 import psym.valuesummary.solvers.SolverEngine;
 
 public class SymbolicSearchScheduler extends SearchScheduler {
-  private final TreeMap<Integer, ProtocolState> depthToProtocolState = new TreeMap<>();
+  private transient final TreeMap<Integer, ProtocolState> depthToProtocolState = new TreeMap<>();
 
   public SymbolicSearchScheduler(Program p) {
     super(p);
-  }
-
-  public static void cleanup() {
-    SolverEngine.cleanupEngine();
-  }
-
-  @Override
-  public void doSearch() throws TimeoutException {
-    result = "incomplete";
-    SearchLogger.logStartExecution(1, getDepth());
-    initializeSearch();
-    if (PSymGlobal.getConfiguration().getVerbosity() == 0) {
-      printProgressHeader(true);
-    }
-    searchStats.startNewIteration(1, 0);
-    performSearch();
-    summarizeIteration(0);
-  }
-
-  @Override
-  public void resumeSearch() throws InterruptedException {
-    throw new InterruptedException("Not implemented");
-  }
-
-  @Override
-  public void performSearch() throws TimeoutException {
-    while (!isDone()) {
-      printProgress(false);
-      Assert.prop(
-          getDepth() < PSymGlobal.getConfiguration().getMaxStepBound(),
-          "Maximum allowed depth " + PSymGlobal.getConfiguration().getMaxStepBound() + " exceeded",
-          schedule.getLengthCond(schedule.size()));
-      step();
-      checkLiveness(allMachinesHalted);
-    }
-    checkLiveness(Guard.constTrue());
-    Assert.prop(
-        !PSymGlobal.getConfiguration().isFailOnMaxStepBound() || (getDepth() < PSymGlobal.getConfiguration().getMaxStepBound()),
-        "Scheduling steps bound of " + PSymGlobal.getConfiguration().getMaxStepBound() + " reached.",
-        schedule.getLengthCond(schedule.size()));
-    if (done.isTrue()) {
-      searchStats.setIterationCompleted();
+    if (PSymGlobal.getConfiguration().getSchChoiceBound() == 1
+        && PSymGlobal.getConfiguration().getDataChoiceBound() == 1) {
+      throw new RuntimeException(
+              String.format(
+                      "Error: symbolic strategy does not support both schedule and data choice bounds as 1. Use other strategies instead."));
     }
   }
 
   @Override
   protected void step() throws TimeoutException {
+    srcState.clear();
     allMachinesHalted = Guard.constFalse();
 
-    int numStates = 0;
     int numMessages = 0;
     int numMessagesMerged = 0;
     int numMessagesExplored = 0;
+    int numStates = 0;
 
     if (PSymGlobal.getConfiguration().getSymmetryMode() != SymmetryMode.None) {
       PSymGlobal.getSymmetryTracker().mergeAllSymmetryClasses();
+    }
+
+    if (PSymGlobal.getConfiguration().isUseBacktrack()) {
+      storeSrcState();
+      schedule.setSchedulerDepth(getDepth());
+      schedule.setSchedulerChoiceDepth(getChoiceDepth());
+      schedule.setSchedulerState(srcState, machineCounters);
+      schedule.setSchedulerSymmetry();
     }
 
     removeHalted();
@@ -106,6 +80,11 @@ public class SymbolicSearchScheduler extends SearchScheduler {
     }
 
     SolverStats.checkResourceLimits();
+
+    if (PSymGlobal.getConfiguration().isChoiceOrchestrationLearning()) {
+      PSymGlobal.getChoiceLearningStats()
+              .setProgramStateHash(this, PSymGlobal.getConfiguration().getChoiceLearningStateMode(), schedulingChoices);
+    }
 
     Message effect = null;
     List<Message> effects = new ArrayList<>();
@@ -239,32 +218,23 @@ public class SymbolicSearchScheduler extends SearchScheduler {
   }
 
   @Override
-  protected void summarizeIteration(int startDepth) {
-    printProgress(false);
-  }
-
-  @Override
-  protected void recordResult(SearchStats.TotalStats totalStats) {
-    result = "";
-    if (totalStats.isCompleted()) {
-      result += "correct for any depth";
-    } else {
-      int safeDepth = PSymGlobal.getConfiguration().getMaxStepBound();
-      if (totalStats.getDepthStats().getDepth() < safeDepth) {
-        safeDepth = totalStats.getDepthStats().getDepth();
-      }
-      result += "correct up to step " + safeDepth;
-    }
-  }
-
-  @Override
   protected void printCurrentStatus(double newRuntime) {
-    String str =
-        "--------------------"
-            + String.format("\n    Status after %.2f seconds:", newRuntime)
-            + String.format("\n      Memory:           %.2f MB", MemoryMonitor.getMemSpent())
-            + String.format("\n      Depth:            %d", getDepth());
-    ScratchLogger.log(str);
+    StringBuilder s = new StringBuilder(100);
+
+    s.append("--------------------");
+    s.append(String.format("\n    Status after %.2f seconds:", newRuntime));
+    s.append(String.format("\n      Memory:           %.2f MB", MemoryMonitor.getMemSpent()));
+    s.append(String.format("\n      Depth:            %d", getDepth()));
+
+    if (PSymGlobal.getConfiguration().isIterative()) {
+      s.append(String.format("\n      Progress:         %.12f",
+              PSymGlobal.getCoverage().getEstimatedCoverage(12)));
+      s.append(String.format("\n      Iterations:       %d", (getIter() - getStart_iter())));
+      s.append(String.format("\n      Finished:         %d", getFinishedTasks().size()));
+      s.append(String.format("\n      Remaining:        %d", getTotalNumBacktracks()));
+    }
+
+    ScratchLogger.log(s.toString());
   }
 
   @Override
@@ -273,6 +243,12 @@ public class SymbolicSearchScheduler extends SearchScheduler {
     s.append(StringUtils.center("Time", 11));
     s.append(StringUtils.center("Memory", 9));
     s.append(StringUtils.center("Depth", 7));
+
+    if (PSymGlobal.getConfiguration().isIterative()) {
+      s.append(StringUtils.center("Iteration", 12));
+      s.append(StringUtils.center("Remaining", 24));
+      s.append(StringUtils.center("Progress", 24));
+    }
 
     if (consolePrint) {
       System.out.println(s);
@@ -283,33 +259,56 @@ public class SymbolicSearchScheduler extends SearchScheduler {
 
   @Override
   protected void printProgress(boolean forcePrint) {
-    double newRuntime = TimeMonitor.getInstance().getRuntime();
-    printCurrentStatus(newRuntime);
-    boolean consolePrint = (PSymGlobal.getConfiguration().getVerbosity() == 0);
-    if (consolePrint || forcePrint) {
-      long runtime = (long) (newRuntime * 1000);
-      String runtimeHms =
-          String.format(
-              "%02d:%02d:%02d",
-              TimeUnit.MILLISECONDS.toHours(runtime),
-              TimeUnit.MILLISECONDS.toMinutes(runtime) % TimeUnit.HOURS.toMinutes(1),
-              TimeUnit.MILLISECONDS.toSeconds(runtime) % TimeUnit.MINUTES.toSeconds(1));
+    if (forcePrint
+            || !PSymGlobal.getConfiguration().isIterative()
+            || (TimeMonitor.getInstance().findInterval(getLastReportTime()) > 5)) {
+      setLastReportTime(Instant.now());
+      double newRuntime = TimeMonitor.getInstance().getRuntime();
+      printCurrentStatus(newRuntime);
+      boolean consolePrint = (PSymGlobal.getConfiguration().getVerbosity() == 0);
+      if (consolePrint || forcePrint) {
+        long runtime = (long) (newRuntime * 1000);
+        String runtimeHms =
+                String.format(
+                        "%02d:%02d:%02d",
+                        TimeUnit.MILLISECONDS.toHours(runtime),
+                        TimeUnit.MILLISECONDS.toMinutes(runtime) % TimeUnit.HOURS.toMinutes(1),
+                        TimeUnit.MILLISECONDS.toSeconds(runtime) % TimeUnit.MINUTES.toSeconds(1));
 
-      StringBuilder s = new StringBuilder(100);
-      if (consolePrint) {
-        s.append('\r');
-      } else {
-        PSymLogger.info("--------------------");
-        printProgressHeader(false);
-      }
-      s.append(StringUtils.center(String.format("%s", runtimeHms), 11));
-      s.append(
-          StringUtils.center(String.format("%.1f GB", MemoryMonitor.getMemSpent() / 1024), 9));
-      s.append(StringUtils.center(String.format("%d", getDepth()), 7));
-      if (consolePrint) {
-        System.out.print(s);
-      } else {
-        SearchLogger.log(s.toString());
+        StringBuilder s = new StringBuilder(100);
+        if (consolePrint) {
+          s.append('\r');
+        } else {
+          PSymLogger.info("--------------------");
+          printProgressHeader(false);
+        }
+        s.append(StringUtils.center(String.format("%s", runtimeHms), 11));
+        s.append(
+                StringUtils.center(String.format("%.1f GB", MemoryMonitor.getMemSpent() / 1024), 9));
+        s.append(StringUtils.center(String.format("%d", getDepth()), 7));
+
+        if (PSymGlobal.getConfiguration().isIterative()) {
+          s.append(StringUtils.center(String.format("%d", (getIter() - getStart_iter())), 12));
+          s.append(
+              StringUtils.center(
+                  String.format(
+                      "%d (%.0f %% data)",
+                      getTotalNumBacktracks(), getTotalDataBacktracksPercent()),
+                  24));
+          s.append(
+              StringUtils.center(
+                  String.format(
+                      "%.12f (%s)",
+                      PSymGlobal.getCoverage().getEstimatedCoverage(12),
+                      PSymGlobal.getCoverage().getCoverageGoalAchieved()),
+                  24));
+        }
+
+        if (consolePrint) {
+          System.out.print(s);
+        } else {
+          SearchLogger.log(s.toString());
+        }
       }
     }
   }
@@ -318,7 +317,7 @@ public class SymbolicSearchScheduler extends SearchScheduler {
   public void print_search_stats() {
     SearchStats.TotalStats totalStats = searchStats.getSearchTotal();
     double timeUsed =
-        (Duration.between(TimeMonitor.getInstance().getStart(), Instant.now()).toMillis() / 1000.0);
+            (Duration.between(TimeMonitor.getInstance().getStart(), Instant.now()).toMillis() / 1000.0);
     double memoryUsed = MemoryMonitor.getMemSpent();
 
     super.print_stats(totalStats, timeUsed, memoryUsed);
@@ -358,7 +357,53 @@ public class SymbolicSearchScheduler extends SearchScheduler {
 
   @Override
   public void reportEstimatedCoverage() {
-    throw new NotImplementedException();
+    PSymGlobal.getCoverage().reportChoiceCoverage();
+
+    BigDecimal coverage = PSymGlobal.getCoverage().getEstimatedCoverage(22);
+    assert (coverage.compareTo(BigDecimal.ONE) <= 0) : "Error in progress estimation";
+
+    String coverageGoalAchieved = PSymGlobal.getCoverage().getCoverageGoalAchieved();
+    if (isFinalResult && PSymGlobal.getResult().equals("correct for any depth")) {
+      PSymGlobal.getCoverage();
+      coverageGoalAchieved = CoverageStats.getMaxCoverageGoal();
+    }
+
+    StatWriter.log("progress", String.format("%.22f", coverage));
+    StatWriter.log("coverage-achieved", String.format("%s", coverageGoalAchieved));
+
+    if (PSymGlobal.getConfiguration().isIterative()) {
+      SearchLogger.log(
+              String.format(
+                      "Progress Guarantee       %.12f", PSymGlobal.getCoverage().getEstimatedCoverage(12)));
+      SearchLogger.log(String.format("Coverage Goal Achieved   %s", coverageGoalAchieved));
+    }
+  }
+
+  @Override
+  protected void reset_stats() {
+    searchStats.reset_stats();
+    PSymGlobal.getCoverage().resetCoverage();
+    if (PSymGlobal.getConfiguration().isChoiceOrchestrationLearning()) {
+      PSymGlobal.getChoiceLearningStats()
+              .setProgramStateHash(this, PSymGlobal.getConfiguration().getChoiceLearningStateMode(), null);
+    }
+    searchStats.reset_stats();
+  }
+
+  /** Reset scheduler state */
+  @Override protected void reset() {
+    super.reset();
+    depthToProtocolState.clear();
+  }
+
+  /** Restore scheduler state */
+  @Override protected void restore(int d, int cd) {
+    super.restore(d, cd);
+    depthToProtocolState.clear();
+  }
+
+  public static void cleanup() {
+    SolverEngine.cleanupEngine();
   }
 
   public static class ProtocolState {
